requirements.txt
fastapi>=0.115.0
numpy>=1.26.0
pydantic>=2.7.0
pyyaml>=6.0.1
sentence-transformers>=3.0.0
uvicorn>=0.30.0
pytest>=8.0.0
